Consolidate overlapping examples (fixes #35)
18 -> 15 examples, each demonstrating one distinct thing:
- Deleted geojson_deferred_visibility_demo.py and
  hexagon_deferred_load_demo.py (all_layers_deferred_visibility_demo
  stays as the deferred-load reference)
- Deleted path_arrows_time_slider_demo.py (time slider covered by
  time_slider_demo; arrows by directed_path_demo)
- New examples/_common.py with the shared GPS track, haversine/speed
  coloring helpers, and an OSM TileLayer shortcut; the two path demos
  now import it instead of duplicating ~40 lines each
- New examples/README.md feature -> example index table
- Doc references to the removed examples repointed
- Bug found by the new import-smoke: basic_usage.py passed camelCase
  initialViewState to the snake_case wrapper and crashed on import -
  fixed to initial_view_state

All 15 examples import cleanly (performance_test_demo skips without
the optional h3 extra, by design).
