+++ examples/directed_path_demo.py
@@ -15,50 +15,11 @@ Usage:
 """
 from dash import Dash, html, callback, Output, Input
 import json
-import math
 
 from deckgl_dash import DeckGL, compute_bounds
 from deckgl_dash.layers import TileLayer, PathLayer
 
-# An ordered track (lon, lat, t_seconds) winding through San Francisco.
-# The hop from index 6 -> 7 is an "impossible" jump (too far for the elapsed time).
-TRACK = [
-    (-122.470, 37.760, 0),
-    (-122.455, 37.762, 120),
-    (-122.440, 37.765, 240),
-    (-122.425, 37.768, 360),
-    (-122.410, 37.772, 480),
-    (-122.400, 37.776, 600),
-    (-122.392, 37.781, 720),
-    (-122.360, 37.795, 735),   # <-- impossible: ~3 km in 15 s
-    (-122.352, 37.800, 855),
-    (-122.345, 37.806, 975),
-    (-122.340, 37.812, 1095),
-]
-
-IMPOSSIBLE_SPEED_MPS = 100.0
-NORMAL_COLOR = [30, 110, 230]     # blue
-IMPOSSIBLE_COLOR = [230, 30, 30]  # red
-
-
-def _haversine_m(lon1, lat1, lon2, lat2):
-    r = 6371000.0
-    p1, p2 = math.radians(lat1), math.radians(lat2)
-    dphi = math.radians(lat2 - lat1)
-    dlmb = math.radians(lon2 - lon1)
-    a = math.sin(dphi / 2) ** 2 + math.cos(p1) * math.cos(p2) * math.sin(dlmb / 2) ** 2
-    return 2 * r * math.asin(math.sqrt(a))
-
-
-def speeds_to_colors(track):
-    """One [r,g,b] per segment: red where the implied speed is impossible, else blue."""
-    colors = []
-    for (lon1, lat1, t1), (lon2, lat2, t2) in zip(track, track[1:]):
-        dt = max(t2 - t1, 1e-6)
-        speed = _haversine_m(lon1, lat1, lon2, lat2) / dt
-        colors.append(IMPOSSIBLE_COLOR if speed > IMPOSSIBLE_SPEED_MPS else NORMAL_COLOR)
-    return colors
-
+from _common import TRACK, speeds_to_colors
 
 TRACK_DATA = [{
     'name': 'Asset Alpha',
